Clean code and tests
